feat/box: base python installation
Unfortunately there is a bizarre issue with Ubuntu where the Python
package is randomly unavailable:
E: Package 'python' has no installation candidate

For this reason I've generated a base box with a Python installation on
it. When you generate the box, *that* box is also subject to the failure,
but you just keep running the process to generate it.