fix/role: install bzip2
It turns out the AWS image for CentOS doesn't have bzip2 installed by
default. This isn't the case for the official CentOS Vagrant box where I
tested if a base installation had this package installed, and that led
me in the wrong direction.

After installing the bzip2 package, the additional customisation isn't
needed.