feat/env: slave for safe vault
Extends the Docker slave generation process to include pulling in
a build container for the Safe Vault project. This also defines the
slave as a 'cloud' for use with EC2.

The ID for the slave has intentionally been left blank because the slave
hasn't been generated yet. This will be supplied after it's generated.