feat/role: wireguard-keepalive
This is a tiny little role that sets up a cron job on a WireGuard
client. It will run once a minute to ping the remote endpoint. This is
necessary for automatically reconnecting an internal slave to Jenkins.
For some reason you need to perform a ping to 'kick start' it again
after it has been disconnected (which it will be during a backup of
Jenkins).

This also extends the backup process to wait on Jenkins to start up and
then reconnect the macOS slave.