chore/slave: increase macOS executors to 3
Due to mobile stuff now being built in Jenkins, we have lots of jobs that
need to run on macOS. Not sure what the effect will be of enabling this
extra executor, in terms of whether it would slow the machine down too
much and we wouldn't be able to see much time improvement from just
using 2 executors.

Also changes the time of the Jenkins backup job to 1AM, since we are
planning on running a bunch of nightly builds at 2AM, so we obviously
want that to be done first.