feat/slave: new macOS slave
This will connect our 2nd macOS slave to Jenkins. Unfortunately, the
original code wasn't really written to accomodate having more than 1
slave connected, so there was some refactoring here to rename the
original variables to include '001'. The 2nd slave is also an identical
set of variables, just using '002' instead.

There seemed to be some kind of issue with both of the macOS slaves
having the same external IP address. I didn't realise it, but WireGuard
actually modifies the config file on the haproxy machine and puts the
macOS external IP address in, and there's also a port reference. I have
no idea how WireGuard is able to communicate with the slaves on these
ports, because I didn't open any on the slave end. We changed the new
slave to have another IP address, but I don't know if it was this that
solved the issue, because after we done that, I also noticed another
mistake.

The provisioning mechanism for the macOS slave has changed. Rather than
passing the IP addresses around, I've created internal and external
inventory files, and the script will use either of these files depending
on whether you're provisioning in the office or not.

The jenkins.yml was updated to provide the 2nd slave. It's worth noting
here that a loop over the osx_slaves group isn't being used because we
have to refer to completely different variables.

This also upgrades the let's encrypt production and staging servers to
use the next version of their API. The staging environment has already
moved to the next version and production will switch soon. See here for
details:
https://community.letsencrypt.org/t/end-of-life-plan-for-acmev1/88430/2