refactor/slave: remove internal/external hosts
Since our 2 slaves are on different internal networks, we won't actually
be able to run the provisioning targets from inside the office. We can
just use the external IPs and remove the need for any additional
maintenance.

This also fixes a little issue with the WireGuard configuration on the
haproxy machine. Adding a new peer involves restarting WireGuard, and in
that case we need to set `SaveConfig` to false, otherwise WireGuard will
overwrite our new configuration when it's shut down.