fix/slave: delete root volume on termination
Not having this setting applied was causing EBS costs to balloon again.
Any slave that was launched was having its 8gb root volume retained
after it was terminated.

This also extends the disk size of the Docker storage volume on the
slave AMI because when you have a configuration with a lot of containers
(like safe-cli) it can sometimes run out of space.